Initialisation

BFlux.BNNInitialiser — Type

abstract type BNNInitialiser end

To initialise BNNs, BNNInitialisers are used. These must be callable and must return three vectors

θnet the initial values for the network parameters

θhyper the initial values for any hyperparameters introduced by the NetworkPrior

θlike the initial values for any extra parameters introduced by the likelihood

Implementation

Every BNNInitialiser must be callable and must return the above three things:

(init::BNNInitialiser)(rng::AbstractRNG) -> (θnet, θhyper, θlike)

BFlux.InitialiseAllSame — Type

InitialiseAllSame(dist::D, like::BNNLikelihood, prior::NetworkPrior)

Initialise all values by drawing from dist