Feedforward Likelihoods

BFlux.FeedforwardNormal — Type

FeedforwardNormal(nc::NetConstructor{T, F}, prior_σ::D) where {T, F, D<:Distribution}

Use a Gaussian/Normal likelihood for a Feedforward architecture with a single output.

Assumes is a single output. Thus, the last layer must have output size one.

Arguments

nc: obtained using destruct

prior_σ: a prior distribution for the standard deviation

BFlux.FeedforwardTDist — Type

FeedforwardTDist(nc::NetConstructor{T, F}, prior_σ::D, ν::T) where {T, F, D}

Use a Student-T likelihood for a Feedforward architecture with a single output and known degress of freedom.

Assumes is a single output. Thus, the last layer must have output size one.

Arguments

nc: obtained using destruct

prior_σ: a prior distribution for the standard deviation

ν: degrees of freedom