Prior and Posterior Predictive

BFlux.sample_prior_predictive — Function

sample_prior_predictive(bnn::BNN, predict::Function, n::Int = 1;

Samples from the prior predictive.

Arguments

bnn a BNN

predict a function taking a network and returning a vector of predictions

n number of samples

Optional Arguments

rng a RNG

BFlux.get_posterior_networks — Function

get_posterior_networks(bnn::BNN, ch::AbstractMatrix{T}) where {T}

Get the networks corresponding to posterior draws.

Arguments

bnn a BNN

ch A Matrix of draws (columns are θ)

BFlux.sample_posterior_predict — Function

sample_posterior_predict(bnn::BNN, ch::AbstractMatrix{T}; x = bnn.x)

Sample from the posterior predictive distribution.

Arguments

bnn: a Bayesian Neural Network

ch: draws from the posterior. These should be either obtained using mcmc or bbb

x: explanatory variables. Default is to use the training data.